+++ CqlSharp/Logging/LogLevel.cs
@@ -1,0 +1,17 @@
+﻿
+namespace CqlSharp.Logging
+{
+    /// <summary>
+    /// Available log levels
+    /// </summary>
+    public enum LogLevel
+    {
+        Verbose = 0,
+        Query = 1,
+        Info = 2,
+        Warning = 3,
+        Error = 4,
+        Critical = 5,
+        None = 6
+    }
+}