+++ CqlSharpTest/SerializationTest.cs
@@ -28,7 +28,7 @@ namespace CqlSharpTest
     [TestClass]
     public class SerializationTest
     {
-        private const string ConnectionString = "server=localhost;throttle=100;ConnectionStrategy=Exclusive";
+        private const string ConnectionString = "server=localhost;throttle=100;ConnectionStrategy=Exclusive;loggerfactory=debug";
 
         private const string CreateKsCql =
             @"CREATE KEYSPACE Test WITH replication = {'class': 'SimpleStrategy', 'replication_factor' : 1} and durable_writes = 'false';";